Handle newlines at beginning and end of selection.
This commit fixes a bug in reflow-selection when the selection begins on the
line before the start of a paragraph, or ends on the line immediately after the
paragraph. In that case, a space was being inserted at the start of the initial
paragraph's first line, and at the end of the final paragraph's last line (and
the blank line after the final paragraph was being lost).