Improve reflow by implementing Minimum Raggedness reflow
This saves me from having to manually fix up result after reflowing.

Before this change, we greedily put as many words as possible on each
line. Given a mix of long words (like URLs) and short words, that can
sometimes render unwanted results.

With this change in place, we add an option to try to avoid short lines
when reflowing. This gives a better (IMO) result that I don't have to
manually fix up.

Code and inspiration from http://xxyxyz.org/line-breaking, this is a
port of the "Shortest path" implementation.